# pkg/nukibridge/api/model_auth.go
/*
 * Keyturner api
 *
 * Keyturner api
 *
 * API version: 1.0.0
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package api

type Auth struct {

	Token string `json:"token,omitempty"`

	Success bool `json:"success,omitempty"`
}