# pkg/nukibridge/api/model_bridge_config.go
/*
 * Keyturner api
 *
 * Keyturner api
 *
 * API version: 1.0.0
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package api

type BridgeConfig struct {

	PairingEnabled *bool `json:"pairingEnabled,omitempty"`
}