# pkg/nukibridge/api/model_callback_object.go
/*
 * Keyturner api
 *
 * Keyturner api
 *
 * API version: 1.0.0
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package api

type CallbackObject struct {

	NukiId int32 `json:"nukiId,omitempty"`

	DeviceType int32 `json:"deviceType,omitempty"`

	Mode int32 `json:"mode,omitempty"`

	State int32 `json:"state,omitempty"`

	StateName string `json:"stateName,omitempty"`

	BatteryCritical bool `json:"batteryCritical,omitempty"`
}