# pkg/nukibridge/api/model_callbacks.go
/*
 * Keyturner api
 *
 * Keyturner api
 *
 * API version: 1.0.0
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package api

type Callbacks struct {

	Callbacks []Callback `json:"callbacks,omitempty"`
}