# pkg/nukibridge/api/model_info.go
/*
 * Keyturner api
 *
 * Keyturner api
 *
 * API version: 1.0.0
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package api

type Info struct {

	BridgeType int32 `json:"bridgeType,omitempty"`

	Ids InfoIds `json:"ids,omitempty"`

	Versions InfoVersions `json:"versions,omitempty"`

	Uptime int32 `json:"uptime,omitempty"`

	CurrentTime string `json:"currentTime,omitempty"`

	ServerConnected bool `json:"serverConnected,omitempty"`

	ScanResults []ScanResult `json:"scanResults,omitempty"`
}