# pkg/nukibridge/api/model_lock.go
/*
 * Keyturner api
 *
 * Keyturner api
 *
 * API version: 1.0.0
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package api

type Lock struct {

	Id *string `json:"id,omitempty"`

	Address *string `json:"address,omitempty"`

	Name *string `json:"name,omitempty"`

	Pin *int32 `json:"pin,omitempty"`
}