# pkg/nukibridge/api/model_lock_config.go
/*
 * Keyturner api
 *
 * Keyturner api
 *
 * API version: 1.0.0
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package api

type LockConfig struct {

	NukiId *string `json:"nukiId,omitempty"`

	Name *string `json:"name,omitempty"`

	Latitute *float32 `json:"latitute,omitempty"`

	Longitute *float32 `json:"longitute,omitempty"`

	AutoUnlatch *bool `json:"autoUnlatch,omitempty"`

	PairingEnabled *bool `json:"pairingEnabled,omitempty"`

	ButtonEnabled *bool `json:"buttonEnabled,omitempty"`

	LedEnabled *bool `json:"ledEnabled,omitempty"`

	LedBrightness *int32 `json:"ledBrightness,omitempty"`

	TimezoneOffset *int32 `json:"timezoneOffset,omitempty"`

	DstMode *bool `json:"dstMode,omitempty"`

	HasFob *bool `json:"hasFob,omitempty"`

	FobAction1 *int32 `json:"fobAction1,omitempty"`

	FobAction2 *int32 `json:"fobAction2,omitempty"`

	FobAction3 *int32 `json:"fobAction3,omitempty"`

	SingleLock *bool `json:"singleLock,omitempty"`

	AdvertisingMode *int32 `json:"advertisingMode,omitempty"`

	HasKeypad *bool `json:"hasKeypad,omitempty"`

	FirmwareVersion *string `json:"firmwareVersion,omitempty"`

	HardwareRevision *string `json:"hardwareRevision,omitempty"`

	HomeKitStatus *int32 `json:"homeKitStatus,omitempty"`

	TimezoneId *int32 `json:"timezoneId,omitempty"`
}