# pkg/nukibridge/api/model_lock_state.go
/*
 * Keyturner api
 *
 * Keyturner api
 *
 * API version: 1.0.0
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package api

type LockState struct {

	NukiState int32 `json:"nukiState,omitempty"`

	LockState int32 `json:"lockState,omitempty"`

	Trigger int32 `json:"trigger,omitempty"`

	CurrentTime string `json:"currentTime,omitempty"`

	TimezoneOffset float32 `json:"timezoneOffset,omitempty"`

	CriticalBatteryState bool `json:"criticalBatteryState,omitempty"`

	ConfigUpdateCount int32 `json:"configUpdateCount,omitempty"`

	LocknGoTimer bool `json:"locknGoTimer,omitempty"`

	LastLockAction int32 `json:"lastLockAction,omitempty"`

	LastLockActionTrigger int32 `json:"lastLockActionTrigger,omitempty"`

	LastLockActionCompletionStatus int32 `json:"lastLockActionCompletionStatus,omitempty"`

	DoorSensorState int32 `json:"doorSensorState,omitempty"`
}