# pkg/nukibridge/api/model_log_entry.go
/*
 * Keyturner api
 *
 * Keyturner api
 *
 * API version: 1.0.0
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package api

type LogEntry struct {

	Index int32 `json:"index,omitempty"`

	Timestamp string `json:"timestamp,omitempty"`

	AuthId string `json:"authId,omitempty"`

	Name string `json:"name,omitempty"`

	Type int32 `json:"type,omitempty"`

	Details map[string]interface{} `json:"details,omitempty"`
}