# pkg/nukibridge/api/model_nuki_lock.go
/*
 * Keyturner api
 *
 * Keyturner api
 *
 * API version: 1.0.0
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package api

type NukiLock struct {

	NukiId int32 `json:"nukiId,omitempty"`

	Name string `json:"name,omitempty"`

	LastKnownState LastLockState `json:"lastKnownState,omitempty"`
}