# pkg/nukibridge/api/model_nuki_lock_state.go
/*
 * Keyturner api
 *
 * Keyturner api
 *
 * API version: 1.0.0
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package api

type NukiLockState struct {

	State int32 `json:"state,omitempty"`

	StateName string `json:"stateName,omitempty"`

	BatteryCritical bool `json:"batteryCritical,omitempty"`

	Success bool `json:"success,omitempty"`
}