# pkg/nukibridge/api/model_scan_result.go
/*
 * Keyturner api
 *
 * Keyturner api
 *
 * API version: 1.0.0
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package api

type ScanResult struct {

	NukiId int32 `json:"nukiId,omitempty"`

	Name string `json:"name,omitempty"`

	Rssi int32 `json:"rssi,omitempty"`

	Paired bool `json:"paired,omitempty"`
}