# pkg/nukibridge/api/model_simple_response.go
/*
 * Keyturner api
 *
 * Keyturner api
 *
 * API version: 1.0.0
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package api

type SimpleResponse struct {

	Success bool `json:"success,omitempty"`

	Message string `json:"message,omitempty"`
}